for tomcat win server 2012 r2

+++ src/main/java/by/repository/UserRepository.java
@@ -2,13 +2,16 @@ package by.repository;
 
 import by.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
 
 /**
  * Created by albertchubakov on 14.08.17.
  */
-public interface UserRepository extends JpaRepository<User, Long> {
+
+public interface UserRepository extends JpaRepository<User, Integer> {
 
    Optional<User>  findByName(String name);
 }
